feat: use int64 in favor of int

+++ internal/models/category.go
@@ -2,7 +2,7 @@ package models
 
 type CategoryFields struct {
 	Name    string `json:"name" db:"name"`
-	GroupID int    `json:"group_id" db:"group_id"`
+	GroupID int64  `json:"group_id" db:"group_id"`
 }
 
 type Category struct {
